# server/docs/swagger.yaml
components:
  schemas:
...
    handler.JobStatsResponse:
      properties:
        available:
          type: integer
        cancelled:
          type: integer
        completed:
          type: integer
        discarded:
          type: integer
        retryable:
          type: integer
        running:
          type: integer
        scheduled:
          type: integer
      type: object
    handler.QueueErrorSampleDTO:
      properties:
        attempt:
          type: integer
        attempted_at:
          type: string
        created_at:
          type: string
        finalized_at:
          type: string
        job_id:
          type: integer
        kind:
          type: string
        last_error:
          type: string
        max_attempts:
          type: integer
        scheduled_at:
          type: string
        state:
          type: string
      type: object
    handler.QueueSummaryDTO:
      properties:
        attention_jobs:
          type: integer
        average_latency_ms:
          type: integer
        average_runtime_ms:
          type: integer
        error_samples:
          items:
            $ref: '#/components/schemas/handler.QueueErrorSampleDTO'
          type: array
          uniqueItems: false
        latest_activity_at:
          type: string
        name:
          type: string
        oldest_remaining_at:
          type: string
        processed_jobs:
          type: integer
        remaining_jobs:
          type: integer
        running_jobs:
          type: integer
        total_jobs:
          type: integer
      type: object
    handler.QueueSummaryResponse:
      properties:
        generated_at:
          type: string
        queues:
          items:
            $ref: '#/components/schemas/handler.QueueSummaryDTO'
          type: array
          uniqueItems: false
      type: object
...
  securitySchemes:
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    name: API Support
    url: http://www.github.com/EdwinZhanCN/Lumilio-Photos
  description: Media management system API with asset features
  license:
    name: GPLv3.0
    url: https://opensource.org/licenses/GPL-3.0
  title: Lumilio-Photos API
  version: "1.0"
openapi: 3.1.0
paths:
  /api/v1/admin/river/queue-summary:
    get:
      description: Get aggregated processing activity per queue, including recent
        error samples
      parameters:
      - description: 'Recent error samples per queue (default: 5, max: 20)'
        in: query
        name: error_limit
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.QueueSummaryResponse'
          description: OK
      summary: Get queue summaries
      tags:
      - Queue
  /api/v1/admin/river/stats:
    get:
      description: Get aggregated statistics about jobs by state
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.JobStatsResponse'
          description: OK
      summary: Get job statistics
      tags:
      - Queue
...
servers:
- url: localhost:6680/api/v1